Running the tests properly
