+++ test/spec/basement/search/search_service.js
@@ -1,11 +1,12 @@
 "use strict";
 
+
 describe("search service", function () {
   var search;
   var $httpBackend;
   var apiToQuery;
 
-  beforeEach(module("gonevisDash"));
+  beforeEach(angular.mock.module("gonevisDash"));
 
   beforeEach(inject(function (_Search_, _$httpBackend_, _$resource_) {
     $httpBackend = _$httpBackend_;
